[release] This plugin allows bracketing of a release cycle. This can be then be used in scenarios to create an index of released artifacts.
The general idea is that a project, for example a distribution project, release will bracket the release artifacts of all other projects so that at the end an index can be created over those released artifacts.

Though it would be nicer if this responsibility could be handled on the bnd side, detailed knowledge of the artifact’s (for example GAVs and URLs are required) that are just not available on the bnd side. 

The plugin was designed for a situation where a workspace gets released to a Nexus along with its external dependencies. The distribution project is then used to hold an OSGi index to all these artifacts. However, there are other scenarios where bracketing a release could be useful.

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>