Fix template metadata extraction
When the template has a metatype.xml, we also need
to extract the parameter names from the template
contents in order to ensure that the user can supply
values for those params even when there is no
explicit metatype for them.

Because this was not being done, the testSrcDir
parameter of a template was not seen, and we presented
a the New Java Project wizard page 2 with no
test source folder.

Signed-off-by: Neil Bartlett <njbartlett@gmail.com>