bndrun: Restore resolve method to return a String
Overload with a method that takes a Converter if custom conversion is
necessary. This avoids most users from being exposed to a list of
VersionedClauses and having to supply a Converter.

Also fixed potential NPE in ResolveCommand where a null List could be
returned. Also restored isOk check before displaying output.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>