launcher: Use java.home to find default java executable
This means by default, the ProjectLauncher will use the same java to
launch the project as is running Bnd. Otherwise, you use whatever the
default java on the platform which is now Java 9 on my system unless
JAVA_HOME is configured to point to some other Java. And JAVA_HOME
is unset in Eclipse.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>