[windows] Since we seem to be unable to find the locked files cause I am trying an avoiding strategy. You can now set a property:
-overwritestrategy

	classic            		like today
	gc				   		runs gc + finalisers
	delay		       		delays 10x0.5 sec
	disposable-names   		writes to a disposable name and 
                       		then symbolically links the proper name to 
                       		this disposable name
	windows-only-disposable-names
					    	Same as disposable-names but only 
                        	when on windows

I can’t test this on Windows so I suggest we try it. In the (default) classic mode the code should be identical.

For SMA this is a showstopper for R3.4. Looking at the problem this should be a real hindrance for anybody wanting to use Bndtools in anger on a PC. So I think we need to include it in 3.4


Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>