[embedded] Minimise updates when keeping storage
If -runkeep is specified then restarting an executable
JAR should not update bundles that have not been changed.
Also, bundles should be updated before the framework
is started so that the changes are more atomic. Before
this patch there was a lot of bouncing start/stop.

This patch changes the starting order. It will update the
bundles (update/install/remove) before the framework
is started. It will also not update a bundle when it is
already installed.

To be installed is defined as:

* The uber jar has a SHA for the bundle
* In the system bundle data file there is a
  file .bnd.launcher/<bundle-id> that contains the sha
* The sha’s match

To test this, there is also now a magic exit code 197 that 
will not exit the launcher with System.exit() but it
just returns. 


Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>