[junit] Added a surprisingly small class that makes it possible to do standard JUnit tests with a full blown framework.
The framework runs in the JUnit process and exports
all packages in the test directory + what the test directory
imports. You can add bundles just like you can in
a bnd(run) file. 

There is some simple support to build temporary bundles.
This is very nice because you can use classes from the
test directory, no jarring required.

I really need this many projects. I started to copy
it between projects but it is always a big pain
in the ass to get all the required dependencies right
because it requires the full bnd + repositories used
in the build.




Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>