classpath: Update containers for new projects in one job
Instead of creating a job per new project, we use one job to initialize
the classpath contains for all the new projects. For a workspace where
hundreds of jobs are imported, this is a significant performance gain
since each job must contend for the bndCall reentrant lock. With hundreds
of jobs, then resulting contention bogs the system down.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>