classpath: Refresh files inline rather than creating a new job
Instead of creating a refresh job per classpath container, we refresh
the files inline. For a workspace where there are hundreds of projects,
this is a significant performance gain since each refresh action may
trigger a classpath container update which would then contend for the
bndCall reentrant lock. With hundreds of jobs, then resulting contention
bogs the system down.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>