build: Refresh build files synchronously
This is necessary since code later in the build depends upon the built
artifacts having been refreshed into the eclipse workspace. Also,
using jobs for this creates too many jobs when there are many projects
and provides no predictability on when then built archives are ready
in the eclipse workspace.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>