builder: Move updating of bnd classpath container before java builder
This is a big performance win since we now update the bnd classpath
container before the java builder runs which means the java code is not
compiled with an invalid class path and thus needs to be rebuild after
we update the classpath in the bnd builder.

We use a CompilationParticipant to get "in front of" the java builder
and update the bnd classpath container if necessary before the java
builder runs. This allows the jars built by dependency projects to be
immediately available to dependent project's java builder.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>