Move Annotation to using the Converter
This change ensures that annotation defaults are honoured by the proxied annotations, meaning that they behave the way that you would expect if they were loaded normally. As a result the bnd code can rely on defaults from the annotations, but it does need to be careful to check the Annotation value if it cares about the property being unset. This change should significantly simplify the process of working with annotations in bnd.

Signed-off-by: Tim Ward <timothyjward@apache.org>