Support the path to local Maven repository to be a symbolic link
When canonical path to the local Maven repository is a symlink, so
that:

    $ file ~/.m2/repository

points either to an existing directory:

    /home/user/.m2/repository: symbolic link to /home/user/.cache/m2/repository/

or a missing one:

    /home/user/.m2/repository: broken symbolic link to /nowhere

target directory should be used (if exists) or created (if
missing). Instead, in both cases MavenRepository fails with:

    java.nio.file.FileAlreadyExistsException: /home/user/.m2/repository
        at sun.nio.fs.UnixException.translateToIOException(UnixException.java:88)
        at sun.nio.fs.UnixException.rethrowAsIOException(UnixException.java:102)
        at sun.nio.fs.UnixException.rethrowAsIOException(UnixException.java:107)
        at sun.nio.fs.UnixFileSystemProvider.createDirectory(UnixFileSystemProvider.java:384)
        at java.nio.file.Files.createDirectory(Files.java:674)
        at java.nio.file.Files.createAndCheckIsDirectory(Files.java:781)
        at java.nio.file.Files.createDirectories(Files.java:727)
        at aQute.lib.io.IO.mkdirs(IO.java:753)
        at aQute.lib.io.IO.mkdirs(IO.java:749)
        at aQute.maven.provider.MavenRepository.<init>(MavenRepository.java:63)
        at aQute.bnd.repository.maven.provider.MavenBndRepository.init(MavenBndRepository.java:494)
        at aQute.bnd.repository.maven.provider.MavenBndRepository.versions(MavenBndRepository.java:462)
        at aQute.bnd.build.Project.getBundle(Project.java:1151)

In the first place, aQute.lib.io.IO's mkdirs method unconditionally
tries to create the directory. Later on, Java NIO Files'
createDirectory method fails as it explicitly doesn't allow the final
directory itself to be a symbolic link:

    try {
	createDirectory(dir, attrs);
    } catch (FileAlreadyExistsException x) {
	if (!isDirectory(dir, LinkOption.NOFOLLOW_LINKS))
	    throw x;
    }

Instead of handling this case in MavenRepository, push the fix down
the call chain, to IO's mkdirs method, so that every caller of IO's
mkdirs method will benefit - if the directory to create is a symbolic
link, create target directory symbolic link points to.

Signed-off-by: Mykola Nikishov <mn@mn.com.ua>