resource: Fix file locking when using jar:file: URL
On Windows, using a jar:file: URL locks the file. This is for 2 reasons.
One, unless setUseCaches(false) is set, the JarURLConnection will hold
the created JarFile open in a cache. Two, even then, the JarURLConnection
holds a FileURLConnection keeps an open input stream on the file and
there is no way to close this inputstream.

So we completely avoid using jar:file: URLs by inspecting the URL and
using either a FileResource or a ZipResource to read the item described
by the URL.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>