gradle: Set jar.archiveName to name of first generated artifact
If we let the archiveName default, then setting project.version can
cause gradle to add a file, which Bnd will not generate, to the jar task
outputs. This can cause gradle to think outputs have changed when they
have not. So we set jar.archiveName so that it will not default to the
name of an artifact that Bnd may not generate.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>