[indexer] Failing test for _remote.repositories with no repo name
This issue is seen from time to time, and forces a user to delete the artifact from their local .m2. The problem seems to be a missing repository name in the aether metadata file. We have no idea how this occurs, but we should work around it if at all possible.

Signed-off-by: Tim Ward <timothyjward@apache.org>