Don't duplicate properties in standalone Workspace
There does not seem to be a good reason to copy the content of the Run file into
the standalone "Workspace". When we do this, the properties are duplicated at
both levels.

As a result it becomes impossible to remove a key from a BndEditModel... in
`BndEditModel#loadFrom(InputStream)` we do a "clear" which involves repopulating
the map from the content of the workspace, but in the standalone case this just
reloads from whatever the content of the file was when first opened. So you
cannot delete a key from the map.

Signed-off-by: Neil Bartlett <njbartlett@gmail.com>