analyzer: Don't use provider import range for self implemented interface
When a class in a package implements a provider type interface in the
same package, exporting that package must not result in a provider import
range for the substitutable import. So we don't consider interfaces when
they are implemented by a class in the same package as triggering
provider range import.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>