resource: Avoid URI resolve which fails badly with UNC names on Windows
We now use a URI object to decode the string and toAbsolutePath to
ensure the path is not relative.

Fixes https://github.com/bndtools/bndtools/issues/1834

This should solve the problem in this bug since the home UNC folder is
no longer referenced since we are not resolving to a URI of that folder.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>