central: Map file to shortest path when multiple paths are present
If the Eclipse workspace is constructed where the same file on the file
system is mapped into multiple Eclipse projects, we need to select one
of the resources. There is no way to know the proper answer, but we
choose the resource whose IPath has the smallest segment count. When
mapping the IPath to an IProject, this will return the project "closest"
to the resource.

Fixes https://github.com/bndtools/bndtools/issues/1847

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>