gradle: Script to publish binary plugin to Gradle Plugin Portal
We download the binaries from Maven Central and then populate the project
artifacts with these binaries. These are then published to the Gradle
Plugin Portal as if they were built by the project.

This publish script needs to be executed _after_ each release is
available in Maven Central so that the same binaries are also available
from the Gradle Plugin Portal.

Fixes https://github.com/bndtools/bnd/issues/2468

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>