gradle: Update SourceDirectorySet.outputDir handling
In Gradle 4.8, the groovy plugin ignored the groovy.outputDir setting
which caused the Bnd build to fail building the Bnd Gradle plugin. I
opened https://github.com/gradle/gradle/issues/5722. The Gradle developer
indicated that the SourceDirectorySet.outputDir property was perhaps a
mistake and that the compile tasks destinationDir property should be set.

This change sets both the outputDir and destinationDir properties which
appears to work fine. Failing to set the outputDir folder caused test
failures for the Bnd Gradle plugin test cases. When/If Gradle removes
the outputDir property, this code should be updated.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>