[analyser] Handle double exports
The Export-Package manifest is calculated from the
Export-Package header in the bnd file, the -exportcontents header, and the Export annotations in the package-info.java files.

This, however, can cause duplicates. This patch
makes sure that packages listed in the private packages are removed from the packages selected by the Export annotation and the Export-Package overrides any attributes of the annotation.

Fixes #2503

Signed-off-by: Peter Kriens <peter.kriens@aqute.biz>