ds: Avoid duplicate extender requirement for version 1.4+
This change checks if the Component annotation class in the analyzer
is itself annotated with RequireServiceComponentRuntime. If it is, then
that annotation will handle generating the requirement for the DS
extender and the DS annotation processing code will not emit the
extender requirement.

Fixes https://github.com/bndtools/bnd/issues/2545

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>