bnd: Fix runtest command so it works again for OSGi build
For runtests, we create a workspace in the folder specified by
--workspace or we use the current working directory. We also do not
require the folder used to be a proper workspace having cnf/build.bnd
since this is not required to run tests.

This commit also includes a number of improvements to FileTree to make
the API is bit more friendly. Both List and varargs versions of the
methods are provided.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>