[tester] Fragments & Lazy
This patch makes the tester handle the case that the
host of a fragment is lazy. The code did not take
STARTING hosts as active so it though the fragment host
was not alive yet.

There can be no confusion with a hanging activator
since the launcher would have reported that as an error.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>