[io] Home directory and getenv
The calculation of the home directory first looks at %HOME%/$HOME and then it uses the system property user.home. 

On Windows %HOME% contains other environment variables, i.e. it is not expanded. Users had to remove this variable to make it work.

This fix expands macros on windows in %HOME%

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>