workspace: Fix ActivelyClosingClassLoader to allow missing files
The OSGi build was broken because of the change to the plugin class
loader to use the new ActivelyClosingClassLoader since it failed if
a file did not exist. The OSGi build compiles AnalyzerPlugins into
cnf/bin which, of course, does not exist at the start of a clean build.
This change allows for a file to not exist and keeps checking for the
file upon each load request.

This change also fixes the class loader to respect the ordering of files
to the class loader by using LinkedHashMaps to hold the File/Wrapper
mappings.

There are also some other refactorings including removing caching of jar
entry names.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>