[trace] Removed deprecation from the Reporter.trace
I get too frequently caught in the case that I need
to provide a reporter but the code also want to log.
It is then very inconvenient to to have an SLF4J logger
AND a reporter.

The reporter is often necessary because the caller
can then look at the errors and react on them, the
log disappears somewhere and is not always associated
with the command.

So I restored the trace method with the function
that it will log to SLF4J. So now the code can
have one mechanism. The SLF4jReporter can in most
cases be used as a default when the reporter is
optional.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>