[shell] bnd command line shell
It is often hard to debug macros and their interaction
with the commands. This shell provides access to
all macros and all bnd commands while running in the
context of a project, workspace, or bnd.

It is also possible to assign values with

	> foo=3

Type help for more info.

The shell uses jline so there is a history. One day
someone might use the completion facility ...

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>