project: Distinguish between locating and invocation of method
We don't want an invocation exception on the first method to report a
locating exception of the fallback method.

I tried to use MethodHandles to invoke the constructor, but got weird
linkage errors when the handler was loaded by a different class loader
such as in the gradle plugin usage of the TestOSGi task type.

 java.lang.LinkageError: loader constraint violation: when resolving
  method "aQute.tester.plugin.ProjectTesterImpl.<init>(LaQute/bnd/build/Project;)V"
  the class loader (instance of <bootloader>) of the current class,
  java/lang/Object, and the class loader (instance of
  org/gradle/internal/classloader/VisitableURLClassLoader) for the
  method's defining class, aQute/bnd/build/ProjectTester, have different
  Class objects for the type aQute/bnd/build/Project used in the signature

Fixes https://github.com/bndtools/bnd/issues/2311

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>