[remoteworkspace] An RPC interface to the Workspace
Adds an RPC interface to the workspace. Initially 
for the JUnitFramework that uses it to analyze the
project setup for running a test and to build
bundles.

This required a RemoteWorkspace service that
defines the RPC. A client factory is provided
to setup a link to a remote workspace. This 
setup uses a registration of the port number in
./cnf/cache/remotews. Since it is possible that
there are multiple processes using a workspace
this register allows n-registrants.

To make the RemoteWorkspace usable it was necessary
to define serializable abstractions for the parameters
of Builder and a Run. To minimize maintenance these
are processed in Builder and Project

This also required adaption of Attrs and Parameters
to work with normal Map objects.

Workspace was also adapted to always create build.bnd
making it always valid so this test was removed.



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>