[junitframework] JunitFramework that uses remote Ws
The JunitFramework was moved to a separate project
and extended. Main change is now a Builder and
the use of the Remote Workspace to analyze the 
project test setup and build bundles.

This way, the only dependency for a test is this
project. 

Dependencies on the framework and run path are also
not necessary to add to the -buildpath or -testpath.
They are all handled in the setup and have their own
class loader.

We probably want to remove the old Junit framework
classes after polling the users. It was never documented
so the number of users is minimal.

This setup should also be possible to make work
in Maven if it implements the remote workspace.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>