pom: Change strategy to pass dependency information from ProjectBuilder
The ProjectBuilder will now collect the dependency information for the
pom since it has full access to the buildpath and buildpath entry
attribute. This allows the use of `maven-scope` attribute on the
buildpath entry to set the scope of the maven dependency. It also allows
the developer to set `-maven-dependencies` in the bnd file to have
explicit control of the pom's dependencies section.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>