[junitframework] Different strategy for extra packages
The previous strategy did not takes uses constraints
into account, it just exported imports and derived
the export version from the import. However, this
caused nasty runtime problems like class cast and
linkage errors.

So now it properly calculates a bundle made out of
the test direct that exports all its transitive imports. This
sounds a bit too much but I've not been able to
figure out if I can handle one level of imports or not.
Ah, think it is better to be safe than sorry

Lets see.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>