maven: For integration tests, name local repo as `local`
This change solved test failures in the index plugin integration tests
where local urls were forbidden but were placed in the generated index
anyway. This seemed to happen because the local repo setup for the
integration test was named `local.central` when the plugin code which
mapped local files to remote urls expected the name to be `local`. So
making this change to the integration tests caused the plugin to
generate the correct remote urls expected by the test.

I am not sure how this test ever passed before and why it started to
fail for me today ¯\_(ツ)_/¯

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>