[http] Old communications throw errors
Fixes #2710

Removed Jetty and the tests and replace the use of the old URL Connectors with the Http Client.

This is NOT completely compatible. If people use the old way to set passwords they need to update to http client. I could find no way to make it backward compatible with good semantics.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>