bndtools: Remove p2Zip repo task from :build for JDK 11
The old p2 ant task uses an old Equinox which fails to run properly
on JDK 11. So we only include the p2Zip task for JDK 8 which is the
canonical build JDK at the present time. If/when we want to switch to
JDK 11 (or later) as the canonical build JDK, we will need to solve how
to build the p2 repo on that JDK.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>