build: Change dist to use with BndPomRepository
Instead of indexing the local release repo folder and packaging the
index in the generated jar, we now put the GAV information into the
pom file for the artifact. This way BndPomRepository can be used to
create a baseline repo since BndPomRepository will download and index
all the dependencies in the dist pom.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>