[runtime] Runtime bundles
This project introduces two bundles and is likely to grow. The bundles are a gogo command bundle that replaces the default
gogo command bundle and that better leverages
the gogo facilities. It also uses an advanced
table based display technique.

The second bundle is a snapshot bundle. It 
will create a snapshot of the framework when
the framework stops. It also has a gogo
command to make an arbitrary snapshot. An
html file is provided to view the snapshot

The snapshot can be used to get an extensive overview of the state of the framework. It captures the framework (bundle,packages,services), the SCR info, configuration, log, and custom extensions.

They are in one project because they share a common information model. Although the OSGi has many DTO's they often are awkward to use or provide just not enough information. A key thing is often indexing. For example SCR repeats service info as full DTO instead of its id, descriptions have no id, etc. An attempt is made to keep the same field names.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>