[builder] Duplicate Export-Package clause exports second and later package with bundle version #2864
Assuming 1.0.0 in package-info.java/packageinfo in foo.bar, the manifest will add 2 exports. One for the package-info.java version and one for the bundle version. The version augment only handles the first clause and ignores the second clause that will then get the default bundle version.

Make a secondary and further export and error if no version is specified

Fix is reporting an error on later exports with identical package name but not version.

Fixes #2864

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>