[parameters] There shouldn't be "Duplicate name..." warnings with pedantic flag set to true #2803
This message spams the warnings as soon as you set -pedantic: true. It would be nice if you don't have to use -fixupmessages Duplicate name in order to get rid of those warnings as those duplicates seem to be common.

Set the `allowDuplicates` flag in the parameters used for headers that actually allow duplicates. This should remove any warnings for duplicated keys in Export-Package, -export-contents, require-capability and provide-capabality.

Fixes #2803


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>