TemplateLoader: Race condition when loading templates
Because TemplateLoader.findTemplates() returns a collection of Promises, we don't want to release our service reference until after those Promises have finished, as they may still be using internal service state.

This fixes an issue where the list of available templates is sometimes not populated when trying to create a new project.

Signed-off-by: Sean Bright <sean.bright@gmail.com>