[launching] Error: Could not find or load main class aQute.launcher.Launcher #2947
Eclipse has deprecated getClasspath and replaced it with
a getClasspathAndModulePath, see https://bugs.eclipse.org/bugs/show_bug.cgi?id=529435

However, the way they did this is not very backward compatible, see

https://github.com/eclipse/eclipse.jdt.debug/commit/89530b29cb538a73f57f4b32cdf3f543258b9bc6#diff-9010920d00bc21b4c2749643470ff0cfL95

I've overridden getClasspathAndModulepath() and ensure the first String[]
is the current classpath.

I've tested this on an older version and that works. When this is build I will
try this on the latest.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>