index - maven plugin: allow absolute local indices
The maven local-index goal allows to set a baseFile that is used to
relativize the found file locations URLs.
If the baseFile property is missing the directory of the output XML file
is used.

There is currently no option (using the Maven indexer plugin) to disable
the relativize step -- which is supported by the used SimpleIndexer
implementation.

This change allows to set a Maven plugin property to use absolute file
URIs.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>