[ds] Check for duplicate resource names
I noticed that if the same name for the component
is used we fail verification. This patch adds
a unique number to the path in the JAR to make
components have unique names.

This generates an error because the only correct
solution is to fix the names. According to the
spec the names must be unique in a bundle. (At least
the Gogo command assumes it.)



Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>