http: Delay on current thread when using inline executor
If we use recoverWith on the delayed promise while using the inline
executor we end up moving work onto the scheduled executors thread.
Since we have only a limited number of schedule executor threads, we
want to avoid long term use. So when using the inline executor, we wait
for the delay to expire on the current thread. This should keep all work
on the current thread when using the inline executor.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>