http client: add timeout to retry logic
We are still seeing some cases where connections seem to hang. So we add
a minimum 5 min timeout to the connection promise to try and recover.
Since we will "lose" the thread running the callable should the
connection hang, we do not want to lose the callers thread for
synchronous requests. So all requests are now async with synchronous
caller's blocking on the result. This means that even if the connection
hangs and the promise times out, we can return the synchronous caller's
thread.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>