httpclient: Refactor for better handling of TimeoutException
We now set the TimeoutException stack trace to that of the "stuck"
request thread for some better diagnostic information. We also attempt
to close the input stream and delete the file of the TaggedData if there
is one for the request thread.

We also set the temporarily set the thread name to the URL of the
request. This can be quite useful in debugging and thread dumps.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>