build: Switch away from Eclipse p2 repos to our own bintray copy
The Eclipse download site is too unreliable for our use. Even with
significant robustness improvements to HttpClient and P2Repository,
the reliability of download.eclipse.org is bad enough to still fail
our CI builds on a regular basis.

I took the content from the 3 p2 repos we used and merged them into a
single indexed repo on bintray. The index will reduce the number of
files which need to be downloaded which may improve CI build times.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>