io: Handle symlinks to non-existent target folders on Windows
A Windows file system oddity means that when a symlink is created, the
file system has to know if the target is a directory. So when we make
a directory via a symlink, we must recreate the symlink so that Windows
knows that the symlink points to a directory and not a file.

See dwFlags parameter:
https://docs.microsoft.com/en-us/windows/desktop/api/winbase/nf-winbase-createsymboliclinkw#parameters

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>