[remotews] Recreate registry when dir destroyed
If people removed the cache directory from the workspace while eclipse is running then the registry was destroyed. This patch polls if the registry exists and recreates it when it is gone.

Also made Launchpad have its own copy of the run spec and added a multithreaded test. LaunchBuilder can create a Launchpad now that is fully reconnected from the builder. I.e. changes do not affect launched Launchpads


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>