[component] Random DS schema version generation
The DSAnnotations analyzer plugin is a singleton in bnd. However, the min (and later max versions) were kept as instance variables. If the build is ran in parallel then I suspect this caused problems. 

Moved the min/max version to a Version Settings object that is a local parameter and therefore not prone to parallel builds.

bnd is not designed to be used in parallel builds but I suspect when it happens we have a lock. Need to look how this works later.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>