[bndrun] Order the bundles of a resolution in dependency order
Although OSGi does not have ordering because you can
dynamically install/uninstall bundles this patch orders
the -runbundles out of a resolution by looking at their dependencies.
If a bundle depends on another bundle then it is tried to start
after it. If there are circular dependencies then of course
this not work well.

To activate:

   -runorder true

If not set we get a more or less random order

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>