[resolve] Cleanup
There was a growing gap between the Bndtools resolving
and the Gradle/Maven/bnd resolving. This is an attempt
to close this gap.

First, a RunResolution is introduced to hold the result
of a resolve process. It should be used to add methods
that use the result of resolving. Static methods
have been provided to resolve.

A tricky issue is that in Bndtools we resolve against
a BndEditModel while the Bndrun approach in Maven, 
Gradle, and bnd assumes static file system. I actually
found out that Bndtools has a different workspace 
OBR repo from the others sadly. 

I've done my utmost to make sure we're now all using
the same classes and no duplication. Please help.

Last, but not least, this patch also includes topological
sorting of the dependency graph. The RunResolution provides
methods to sort the list of resources by random, name, 
least dependencies first, or last. I suspect that
the leastdependencieslast could be quite effective 
in minimizing the jojo effect. We'll see

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>