[workingsets] Ignore workings with non-java-id name
Some people want to maintain private working sets but
bnd would remove a project from a working set if
the -workingset was used in a project.

This change makes bnd ignore working sets that do
not follow the required naming pattern. Using such
a pattern for a workings therefore implies that the
user is on its own.

For example, a .Private working set will never be 
touched by bnd.

Fixes #3207

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>