[workspaceserver] Added a test bundle
To prepare for the @RunWith support in Launchpad
I need a bundle that consists of exactly the 
project bundle (or one of its sub bundles) together
with the contents of the bin_test dir.

The test classes do not generate imports but handle
their imports via DynamicImport-Package. The
Service-Component header is merged between the original
bundle and test part.


Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>