[launchpad] Support for @RunWith
Launchpad in défault mode works quite well but
it falls apart when you mix classes between the inner
and outer. It seems that a App loader classes can
actually pick up a class from a framework loader
when it is in the API. (This happened with the
Service Component Runtime and the unfortunate
dependency on Promises).

This patch creates a special bundle that is
the project's bundle + the test directory. The
LaunchpadRunner is a JUnit runner that takes
care of all the details.

Signed-off-by: Peter Kriens <Peter.Kriens@aqute.biz>