header: Change duplicate key handling code to avoid O(N^2) calls
If a key is used N times in a header, the code to deduplicate the key
name used O(N^2) containsKey calls to find a deduplicated name. When
large headers are parsed, this can consume considerable time. One
scenario called containsKey 265M times! This fix uses a separate map to
hold the last used deduplicated key name so the next deduplicated key
name can be computed and confirmed available in a single containsKey
call.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>