attr: Fix handling of plain List and avoid regex if no type specified
The previous code used a regex which would not match a plain List type.
So the key ended in `:List` and the value type was String. This fix
correctly handles plain List as the type `List<String>`.

Furthermore, we avoid regex matching when the key does not contain a `:`
which is the most common scenario. This can provide a small performance
boost.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>