build: use new BaselineIgnore annotation instead of -fixupmessages
Since there is only a single error for all baseline issues, using
fixupmessages to elide specific, minor baseline issues is over broad and
can hide more serious baseline issues. So we use the new BaselineIgnore
annotation to ignore baseline issues on specific elements.

For aQute.libg, we change to update the major package version since
aQute.libg is not used as a bundle and also because it is built before
the annotations project and thus cannot use the new BaselineIgnore
annotation.

Signed-off-by: BJ Hargrave <bj@bjhargrave.com>